pyproject.toml
[project]
name = "fastapi-example-api-search"
version = "0.1.0"
description = "Тестовый backend-проект на **FastAPI**, демонстрирующий интеграцию с внешними API через единый интерфейс поиска."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.8",
]

[tool.ruff]
exclude = [
    ".venv",
    ".vscode",
    "temp_*",
    "test_*",
    "for_upgrade"
]
src = ["src"]
line-length = 150
target-version = "py312"

[tool.ruff.lint]
select = ["E","F","UP","B","SIM","I","W"]
ignore = ["E712", "B008"]
